+++ examples/filesharer/_request/list.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+go run ./../../cmd/hls/hls-filesharer/client/hls-filesharer-cli -k localhost:8572 -f Bob -d list -a 0