+++ examples/pinger/_request/ping.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+go run ./../../cmd/hls/hls-pinger/client/hls-pinger-cli -k localhost:7572 -f Bob