+++ examples/remoter/_request/exec.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+go run ./../../cmd/hls/hls-remoter/client/hls-remoter-cli -k localhost:7572 -f Bob