+++ internal/services/filesharer/pkg/app/init_service.go
@@ -29,7 +29,7 @@ func (p *sApp) initExternalServiceHTTP(pCtx context.Context, pHlkClient hlk_clie
 
 	buildSettings := build.GetSettings()
 	p.fExtServiceHTTP = &http.Server{
-		Addr:         p.fConfig.GetAddress(),
+		Addr:         p.fConfig.GetAddress().GetExternal(),
 		Handler:      http.TimeoutHandler(mux, buildSettings.GetHttpHandleTimeout(), "handle timeout"),
 		ReadTimeout:  buildSettings.GetHttpReadTimeout(),
 		WriteTimeout: buildSettings.GetHttpHandleTimeout(),